~time 6m resources.py
